vpn-provider: Interpret VPN route ProtocolFamily as number
The vpnconnection UserRoutes dbus property expects the ProtocolFamily to
be specified, as explained in the documentation:

int ProtocolFamily: Protocol family of the route. Set to 4 if IPv4 and 6
if IPv6 route.

With the current code a GetProperty request will return an integer with
value 4 or 6 as expected. However, when the property is set, it expects
a value of 52 (ASCII char '4') or 54 (ASCII char '6').

This change makes it so that when setting the property, integer values
of 4 or 6 are also accepted without dropping support for '4' and '6'.