[connman] Avoid setting Favorite flag without signal on service create. Fixes JB#48468
When a new service is created using __connman_service_create() the
Favorite flag is set to mirror the AutoConnect flag. However the service
may already exist. If so, any change of the flag value should emit a
dbus signal, so that anything keeping track of the value knows.

In particular the lack of signal causes difficulties for
QNetworkConfigManager and QConnmanEngine in qtbase, which under some
circumstances will be waiting on a Favorite PropertyChange dbus signal
in order to trigger a move to the online state. If the Favorite value is
set to true on creation without a signal, QNetworkConfigManager may stay
offline even though there's a connection.  This can happen with WPA-EAP
networks, for example.