[openfortivpn] Implement VPN plugin for openfortivpn. JB#48797
Co-authored-by: Jussi Laakkonen <jussi.laakkonen@jolla.com>

This adds a new VPN plugin that uses openfortivpn binary to access
FortiNet VPNs with the help of ppp. Code is based on existing L2TP and
OpenConnect plugins. Plugin structure follows same approach as in L2TP
plugin. Property management is similar to OpenConnect plugin.

Additionally, via patch in openfortivpn, --trust-all-certs, is supported
as a property to disable certificate signature checks. Following
properties are supported so far, which are saved to the settings:

 Name                              Value                Option
 openfortivpn.AllowSelfSignedCert  string: true|false   --trust-all-certs
 openfortivpn.TrustedCert          string: fingerprint  --trusted-cert
 openfortivpn.Port                 string: digits

By default, if Port is omitted, openfortivpn uses port 10433.

Signed-off-by: Alexey Andreev <a.andreev@omprussia.ru>
Signed-off-by: Jussi Laakkonen <jussi.laakkonen@jolla.com>