iptables-validate: Rewrite rule syntax parser
New parser is much more consistent, and allows for more variations than
the previous one - multiple -m options are accepted, every option has
a definition if it can accept negation character, or not. The arguments
are verified for consistency with a whitelist of arguments, not with
a combined set of a blacklist + option whitelist.

Some fixes were also introduced to the verification functions, more
edge cases are handled.

Some assumptions were changed, so the rules that pass or fail with this
parser are different. This is done on purpose.