storage: Add support for changing storage directory
[storage] Add support for changing storage directory. JB#47637

This implements support for changing the storage directory for service
settings. User specific WiFi and VPN services are only supported. All
other services will remain in the system main storage directory.

Storage registers ChangeUser D-Bus method to "net.connman.Storage"
D-Bus interface when setup via connmand. And when using for VPN
"net.connman.vpn.Storage" D-Bus interface is used.

Each storage directory change is done when storage D-Bus method
ChangeUser is called with an user id (UID) as a parameter. The UID is
verified with getpwuid() and must be a valid user on the system having
a valid login shell available in /etc/shells (getusershell() is used).
System user (UID = 0) can be changed as well, usually user "root",
effective only if there was another user set, and causes the storage to
return back to original use. In such case storage.c uses the configured
storage root or the default one if it is not configured. Path for user
storage is formed by concatenating user home directory and
DEFAULT_USER_STORAGE which is defined by userstoragedir as a build time
configurable variable in Makefile.am.

Attempts to change to the same user that is set, or setting the main
system user will result in D-Bus error .AlreadyEnabled and nothing is
done. If there is user directory already set and another directory is
beings set the services and VPN providers in the old user directory are
removed after the pre callback is executed. This is done also when
returning to system main directory use.

__connman_storage_set_register_dbus() is called to register the
interface and to setup the callbacks. The callbacks are used for
technology and service removal and loading and are executed in the
following order:
 - pre is the callback that is executed as first to do tech cleanup
 - unload is called to remove the current services from use
 - load is called to load the services/providers with new directory
 - post is executed to restore the newly loaded settings
 - finalize is executed as last when change is done

A new boolean for disabling service removal when changing storage
directory is added into storage.c. This, only_unload, is used to
indicate the storage dir change for other parts of storage.c.

The retrieval of the services is now done in two phases. First, the
system services are requested and this filters out the system WiFi
services if the storage directory was changed. Then the new storage
directory WiFi services are included in the list, if the user specific
new storage directory is set. With VPN providers, the returned list
contains only either of the VPN providers: system providers with default
storage directory and new storage directory, the user VPN providers.