technology: Add pre and post setup functions for user change
[technology] Add pre and post setup functions for user change. JB#47637

When doing preparation setup all technologies in use will be disabled.
All pending scans are disabled, by device and to make sure that each
reply is sent __connman_technology_scan_stopped() is called in addition.
Offline mode is not changed and does not affect the preparation setup.

When doing post setup the offline mode is read from the new settings
file. Next each technology is updated based on its "Enable" value in the
settings file. If the technology is to be enabled but ConnMan is in
offline mode enabling of the technology is skipped and enable persistent
is set to enable restoring the technology when leaving offline mode. To
ensure that overrides are not set the offline mode is set first if it is
disabled and currently in offline mode.

According to __connman_technology_add_device() the power indication is
not done if a non-rfkill device is simply enabled.
__connman_technology_enabled() will trigger the change. This is added to
technology_affect_devices() functionality to ensure that proper power
indication is done when device reports that it is already enabled.