vpn-provider: Add functions to load and unload providers
[vpn-provider] Add functions to load and unload providers. JB#47637

Load providers for each VPN driver separately using
provider_create_all_from_type() when user/path for storage is changed.
Process is identical to when vpn-provider.c is initialized.

Provider unloading uses the array of provider identifiers to delete the
providers. This is to be called only when user is changed and storage.c
has the unload_only boolean set to avoid having the provider files to be
removed as well.