vpn-settings: Implement user override for running a VPN
[vpn-settings] Implement user override for running a VPN. JB#47637

Add possibility to override the user that is used to run a VPN binary.
This setting is not read from configs and is a runtime setting only.
The main use for this is when active user has been changed and the VPN
binary is configured to be ran as different user.

The override is used if the VPN plugin specific config or the VPN daemon
config has an username specified. If the configured username is a system
user or is missing override is not used. Reason is that VPN's without
config or configured system user may have been set because of system
limitations.

For allowing to set the system users that are not overridden by the
active user change setting "SystemBinaryUsers" is implemented. This
string list, similar to SupplementaryGroups, is read from the
DACPrivileges group and can be defined only in the vpnd main config
file. If this setting is omitted, override is prevented only for the
root user or the user running the process (uid equals process euid).