vpn: Add support for setting user dirs via D-Bus
[vpn] Add support for setting user dirs via D-Bus. JB#47637

Reorganize initialization in main.c to have D-Bus and log setup before
storage and rest are initialized. Storage requires D-Bus to be
initialized before the user change D-Bus interface is registered.

Register only unloading and loading of provider functions to storage.c
since with VPNs no preparations or post actions are required. In
addition, use the __vpn_settings_set_binary_user_override() as
finalization function to change the user for running a VPN, if it was
previously set.