[unit] Create tests for storage.c. JB#47637
Test:
 - Basic use of storage.c
 - Global settings
 - D-Bus and user change, with multiple user changes as well
 - Invalid users
 - Service file use is tested, also in user change
 - Callback use in user change is tested
 - Tests for technology.c callbacks are included.

Because the unit test is not running threads the callback counts cannot
be properly checked as the storage.c is normally used in separate
processes. In the test the counts are not representing correct amounts
because other "component" makes changes to certain values. It may not be
feasible to separate all the content for connmand and vpnd specific to
satisfy unit testing needs.

Also verify that root and user files are not removed when changing user
in scenario where the services and providers are not tested for removal.

Added service.c and vpn-provider.c functions are not tested with this unit
test.