storage: Add internal function for changing user
[storage] Add internal function for changing user. JB#49403

Implement __connman_storage_change_user() that is similar to the D-Bus
method function with small differences. Appropriate error is returned
and -EINPROGRESS means that the call to vpnd was made and answer is
pending. The pending message is omitted and this is handled now in
change_user_reply().

This function is to be used by login manager listerners. With callback
the result of the operation will be returned back to the caller, when
D-Bus calls finish.

To support initial setup after storage is initialized but prior to
initializing other components a preparation mechanism for user change is
implemented. This change is to be done by connmand at the early
initialization stages that changes only the user paths to correspond to
the path of the currently logged in active user. No preparations,
unloading, loading or post actions are done because the components that
implement these (technology.c, service.c etc.) are not initialized yet.

The idea is to do the process in reverse when preparing, connman sets
first the paths then communicates with vpnd. The preparation is not
necessary for the vpnd as it can process the D-Bus method call only
after it has been initialized - by systemd. Thus, everything is set when
the user change message is processed.

When preparing the user change callback is called twice, first with
-EINPROGRESS to indicate that change has been made and pending for vpnd
result. Then when vpnd returns it will be called again. If in any of
these steps there is an error it will trigger setting the user back to
user "root", so default storage location is used. In this case the user
change cb is called with uid 0.

And also some other fixes:
  Fix also service and provider counter in system and user getters.
  Replace in these the use of g_realloc() with g_renew(). Use also
  unsigned ints in the getter counters.

  Ignore the already set user VPN dir, this should not even happen as
  the first check on the user storage dir would report the error. Keep
  the debugging message there still.

  Remove unnecessary variable path and use correct one in debug when
  using the internal user change.

  In case the __connman_storage_change_user() fails to change user in
  preparing or the D-Bus fails at any part free the struct
  change_user_data *user_data and relevant conten.

  Handle D-Bus Timeout and TimedOut error messages and set appropriate
  ETIMEDOUT error value. Reduce some explicit handling of the error
  messages to create replies. __connman_error_failed() handles already
  most of them.

  Display error if the uid set using the internal user change has
  greater value than max of unsigned 32bit integer. This is a precaution
  for uid_t being redefined.