[technology] Delay enabling if tech is rfkill enabled or busy. JB#49436
Use delayed enabling if the tech to be enabled is rfkill enabled when
re-enabling in multiuser environment or when the enabling reports being
busy. A technology can be busy because of pending reply, that is being
handled by the timeout function, or because of other component / lower
layer. The delay is set to 300ms and timeout function attempts to enable
until enabled (0 or EALREADY is returned).

For the need to reply to a pending power change request a generic error
handler is implemented. This allows to send a reply to the pending power
request with proper error code and reduce duplicate code. Sending
ETIMEDOUT error in other error cases is not proper way to do it, better
to send ECANCELED when the action is canceled.

Also, use EINTR when canceling scans. This seems to be the correct error
code that is to be passed when canceling request. Do the canceling of
the scan before sending pending reply to power change call in the tech
disabling when user is changed - as it is done in other places in
technology.c.