service: Unload VPN services and always reply from GetProperties.
[service] Unload VPN services and always reply from GetProperties. JB#49504

Unload also the VPN services when changing user in multiuser
environment. This is to cover situations when the user change from vpnd
hasn't yet signaled about removed VPNs.

Always send a reply from GetProperties. This may not be reached but
returning NULL and leaving the pending call awaiting will make D-Bus
daemon unhappy as the default maximum amount of pending call replies is
set to 128.