[storage] Handle manually removed and added services. JB#49436
Add use of unload() and load() callbacks to be able to properly notify
about removed and new services/providers. Removal and adding of the
services can also be manual and in this case the information should be
also forwarded.

When a service is removed it is only required to be unloaded since one
cannot remove what is already deleted. Only the removed service is given
as the sole item in the service list for unload callback. Loading
callbacks should behave in a way that if the identifier is already
present just ignore the service, thus load() should be possible to call
multiple times.

This does not interfere with service/provider removal via connmand/vpnd,
since the removal in service.c and vpn-provider.c also calls
__connman_storage_remove_service() which removes the actual files. After
that the inotify indication about removed ones is received and in then
the subdirs are only cleared for the entry, unload() callbacks should
ignore a missing service. In loading the detection of new services does
another initializtion call but this cannot be avoided since the
functionality for load() is to get all available services.

Move load and post action callback use at the end of the
change_storage_dir(). No need to have them both in the switch.