[technology] Add a generic rfkill-agnostic state changed func. JB#49436
Implement technology_changed_state() to make it easier to indicate a
state change on a technology regardless of the rfkill use. This takes
the functionality from __connman_technology_{enabled,disabled}().

Main reason is to use this when changing the user to get all proper
indications sent via D-Bus. Also when using delayed enabling this is to
be used to make the code less obfuscated.