[technology] Disable all techs regardless of the keyfile. JB#49436
Disable all technologies even when keyfile is missing. Even though
missing keyfile usually means that there shouldn't be any technologies
enabled it can be that the keyfile has been already manually removed or
deleted prior to calling the __connman_technology_disable_all(). Utilize
the return value as a indication of failing to disable a tech instead.