[wifi] Optimize reporting of Wi-Fi networks to connman core. JB#47990
Most of the unnecessary D-Bus traffic between connman and the UI is
generated by low-signal networks that appear and disappear in the
scan results. To add insult to injury, these insignificant changes
in the network list reset the autoscan timer which makes scans to
happen every 5-10 seconds if you're in a busy place.

This commit is an attempt to find a balance between quickly finding
known (pre-configured) networks and yet not flooding the system with
meaningless/unusable information.