[storage] Add generic storage dir getter. JB#49651
The new function allows to get the storage dir for the service using the
identifier. This is required in multiuser environment, as some services
are saved to STORAGEDIR and some are to USER_STORAGEDIR. This function
makes it easier and cleaner to use. Internally, the existing
storagedir_for() is used. If the service identifier is invalid NULL is
returned as storagedir (TODO: should STORAGEDIR be the default instead
of NULL?).