[config] Register to storage notifications for config inotify. JB#49506
Listen and react to user id changes in config.c. Add a notifier for the
new user to listen for config changes. If the user id is different from
the one connmand is running, add notifier to it and copy the storagedir
of that user to be removed later on from inotify (shutdown or another
user change).