[connman] Replace default route opt with split routing use. JB#45606
Default route option "DefaultRoute" is replaced with split routing use.
The use is the complete opposite compared to default route option. When
split routing is set as "true" the VPN is treated as non-default route,
and when set as "false" (the default) VPN is used as the main gateway
for traffic. This change removes the use of default route from connmand.

To make possible to use split routing service.c:set_split_routing() is
changed to __connman_service_set_split_routing() and
provider.c:connman_provider_set_split_routing() using the aforementioned
function is implemented (for plugins).

Cleaned up some code as well (intendations). Removed our default route
connection.c additions which proved to be unnecessary with split
routing.