vpn-provider: Support split routing option for VPN providers
[vpn-provider] Support split routing option for VPN providers. JB#45606

Add the same option as service.c has on connmand side to vpnd in
vpn-provider.c. Replace all uses of default route except in supporting
it as a legacy option when loading from config value and receiving via
D-Bus API. SplitRouting is a boolean in every form, internal, settings
file and in D-Bus API.

The "DefaultRoute" is converted to "SplitRouting" if it exists in the
settings file. Also, if some component wishes to use still
"DefaultRoute" it is converted to "SplitRouting" when received via D-Bus
API. DefaultRoute "true" means the same as SplitRouting bool false.
Property change notifications on DefaultRoute are not sent anymore and
these are replaced with SplitRouting notifications.