technology: Move tech value loading to separate function
[connman] Move tech value loading to separate function. Fixes JB#50601

Created new function based on technology_load() to load only the values
from the keyfile. This is required to be done also when user changes so
it is executed when the configuration for a user is being loaded. Saving
is set to be the last operation to not to clear tethering values in
settings file.

Also, the values for tethering are added to be cleared in the user
change preparation function. This eliminates the leak of other user's
tethering passwords to the next logged in one.