inet: Add function for detecting a default route
[inet] Add function for detecting a default route. JB#51404

Support checking of network route being as default route. If host and
netmask are set as INADDR_ANY then the route is a default route. Gateway
can be unset (NULL) or pointing directly to the gateway IP. Check for
this is, however, omitted because it would need ipconfig to get the set
IP for comparison and ipconfig is different for connmand and vpnd.