vpn-provider: Don't manage routes and react to SplitRouting changes
[vpn-provider] Don't manage routes and react to SplitRouting changes. JB#51404

This change removes the '-r' option to toggle whether to manage routes
or not in connman-vpnd. Remove all functionality from vpn-provider.c
that was used for managing routes. Remove all functions managing the
routes. All routes should be managed by connmand only.

The change was suggested by Daniel Wagner on the mailing list:
https://lists.01.org/hyperkitty/list/connman@lists.01.org/message/
YTSQ636O7GROD4NZH6B4AVQMTO6SN3QO/. The change was originally
introduced in a7851eb04e67a6c046c1665157a97fd6b2d42462 for the most
part.

Implement provider_property_changed() to be registered with every call
to connect_provider() for the signals sent about property changes on
VPN services. React only to "SplitRouting" value changes coming from
the VPN service via connmand.

Drop also the saving of the value to keep it only in one place and to be
mainly controlled by connmand. Allow to load the "DefaultRoute" by
converting it to "SplitRouting" but ignore "SplitRouting" value from
provider settings file. Having this in settings file will print a
warning, and is cleared away at first save done to that VPN.

Also always notify all components listening for the changes over VPN
D-Bus API on SplitRouting changes coming from connmand. The signal will
be reacted upon also by connmand but provider.c will ignore setting of
the same value to avoid D-Bus signal loop. For this the
vpn_provider_add_boolean() is modified to have an option to have th
signal sending forced regardless of the value change. This is required
for notifying the components listening for the changes via VPN D-Bus API
when connmand/connman-vpnd crashes and the value is not written yet to
the VPN service settings file and the component would not get an update,
for example, to "true" value it has if connmand and vpnd have false (by
default, vpnd has the value as false). There are duplicate signals sent
but this approach ensures that all components will know the correct
status of SplitRouting.

Load "DefaultRoute" directly as a boolean, if the value is is omitted or
wrong set "SplitRouting" as false.