provider: Use service move to guarantee proper handling of split route
[provider] Use service move for proper handling of split route. JB#51404

Improve checks and functionality in connman_provider_set_split_routing()
to ensure the proper handling of split routing value changes by
utilizing the separated service moving
(__connman_service_move_service()) function to reorder the services
properly. When a VPN is to be set as split routed it has to be used as a
target, and when VPN is not split routed it should be regarded as the
default service.

If the VPN is not connected only the split routing value is changed for
the service. If there is an error in moving the connected service use
__connman_service_split_routing_changed() to signal vpnd about the
current status of split routing.

To ensure that setting file has the SplitRouting value if changed in VPN
settings by vpnd the vpn_service must be saved. This should be done in
provider.c since in service.c the saving is done explicitly as well
after changing SplitRouting value.