vpn-util: Create utility file for VPN core and plugins
[vpn-util] Create utility file for VPN core and plugins. JB#48841

Create vpn-util.c for VPN core and plugins to use. This is a first step
in establishing some sort of lib for them.

Add functionality to get the user struct passwd and struct group using
the username and groupname. This was previously only within
vpn-settings.c.

Expose is_system_user() as vpn_settings_is_system_user() for plugins to
use.

Add function to create dir for a VPN plugin. The basename of the given
path is used and it must have a prefix of "/var/run/connman-vpn/",
"/var/run/user/" or "/tmp/" and must not contain ".." or "./". One
directory element must be provided after the prefix, e.g. "/tmp/dir/".
If the basename path is an existing dir permissions and ownership is
changed according to the request ones, otherwise the file is removed and
then created, as if it never existed. g_unlink() is used to handle safe
removals of symlink, which also enables detection of parent dir write
permissions - error is returned unless the dir exists, in which case
ownership and permissions are attempted to be set accordingly.

In order to reduce the potential of using the vpn_util_create_path()
with a malicious purpose the prefixes limit access only to run-time and
temporary locations. VPN core and plugins do not need more access to
file system, but to create a temp/run-time dir, e.g., for a pid file.