vpnc: Support setting the pid file path to /var/run/user
[vpnc] Support setting the pid file path to /var/run/user. Fixes JB#48841

Use the new util and settings functions to get the user that is used to
run the VPNC plugin in order to set a correct pid file path. If system
user is used utilize the default path.

Path prefix is set to /var/run/user, and suffix vpnc/pid is added as
well. With user 1000 this results to pid file path of
/var/run/user/1000/vpnc/pid.