[systemd login] Handle ENOENT in active session check. Fixes JB#51399
Handle ENOENT as a report of session not being ready yet (undocumented
in the man pages but kind of obvious from systemd code). In such case
the polling should be kept in main loop since this is reported only when
the system is starting. This change helps to avoid the situation where
systemd_login.c quits too early in attempting to get the active user for
the session. The result would be improper use of the autologin user as
root user, which is not desired.