vpn-settings: Add SystemBinaryUsers conf option and system user check
Add "SystemBinaryUsers" user list to vpnd main config only to allow
defining which users should be regarded as system wide "privileged"
users. The user list is read as a string list from DACPrivileges group.

Add function for checking if the user is a system user, either root,
effective uid of the process or one in the "SystemBinaryUsers":
vpn_settings_is_system_user(). This function can be used by VPN plugins
to check if the user set to run the VPN binary is a regular or system
user since different behavior can be expected for run-time file
creation.