[vpn-provider] Emit conn. created signal if loading form keyfile. Fixes JB#50997
Emit the connection created signal also when loading the VPN connection
from keyfile. This signal was sent in every other case where VPN
connection was created (via D-Bus or from config).

This fixes the issue when an UI component retrieves the VPN list using
the VPN Manager interface before user change has completed. Vpnd loads
VPNs as the last step in the user change process, so the requested list
can be empty but every load of a (new) VPN should be signaled as well.