service: Improve service sorting with more state checks
[service] Improve service sorting with more state checks. JB#51404

Sort states also by the state and when checking for preference use it
for equal states only.

In case the VPN is split routed rely on the order set for the service.
This change makes the split routed VPNs to be arranged in the service
list to be after the transport instead of being after all the online
state services.

Also do some cleanup and allow VPNs to be set to be depending on each
other, the transport VPN must be non-split routed whereas the depending
VPN must be split routed.