iwd: Enable/disable auto_connect on known networks
Change in stragety how ConnMan is driving the auto connect
strategy. Instead relying only on ConnMan doing the auto connect and
disabling all auto connects on the know networks object, we pass
through the autoconnect setting from ConnMan to iwd and rely on iwd to
initiate the initial connect. We have to do this because iwd will
disable scaning if auto connect is disabled.

Luckily ConnMan is able to synchronize itself on state changes from
the external daemons. No internal changes are necessary for ConnMan.

The only changes needed are inside the iwd plugin. We need to store
services' autoconnect status in the iwd_network and iwd_know_network
data structure. This is due the fact know networks are decoupled from
the network objects. So we can't do a reverse lookup. Thus we store
the autoconnect state in iwd_network object and update on any possible
chance the autoconnect variable. If the known_networks' auto_connect
and services' autoconnect do not match do a property change on the
iwd's known_network D-Bus object.