storage: Always call finalize cb, add user change status check
[storage] Always call finalize cb, add user change status check. JB#52114

Move finalize call on connmand storage.c user change to be always
executed. Same uid can be reported back by vpnd also when doing the
first user change in preparing mode after boot/restart. Otherwise
finalize would be ignored when in preparing mode.

Add function connman_storage_user_change_in_progress() for checking if
the user change is ongoing. This uses vpn_change_call, which indicates
the status of the user change directly. When there is a call set it is
pending a reply from vpnd and the value is set to NULL always when the
call ends with success or error.