vpn: Fix memory leaks with PendingCall
According to the D-Bus docs, dbus_pending_call_set_notify "sets a
notification function to be called when the reply is received or the
pending call times out".

Thus, we should not expect any further calls to the registered function
than the one that was triggered. If the idea is wait for a longer
period of time, this can be done with increasing the timeout. Meanwhile,
the possible causes of a timeout is the other party not sending it at
all, which in my testing was possible to be achieved by crashing the
other process.

If we let the callback function just return without unref-ing the
pending call, we get a memory leak.

If vpnd crashes before sending connmand a reply to GetConnections, the
particular vpn_connection data structure will have both connect_pending
set to true, and call set to the pending call that would not receive any
response anymore, and a re-established vpnd process would not attempt
a reconnection.