vpn: Ignore VPN added signals if user change is in progress.
[vpn] Ignore VPN added signals if user change is in progress. JB#52114

Don't react to VPN added signals if user change is still in progress.
This is because storage.c hasn't yet set the paths until the change is
complete and VPN service files cannot be loaded, and existing VPN
service is treated as a new service.

Removed VPN signals can be sent when changing user, as unload callback
on vpnd side is executed during the process.