service: Load and apply service settings after D-Bus registration
Change the service registration order to first register the service to
D-Bus and only after that load the service and apply its variables. This
ensures that the values that should be signaled separately, e.g.,
SplitRouting, can be sent. Otherwise the sending of the signal fails as
the path is not registered.